Split mqtt client tests
